benchpress/requirements.txt
PyYAML==3.12
py-cpuinfo==5.0.0
dmidecode==0.9.0
